App(conf): different types -- unwrapping Config on bootstrap
In one of the previous commit (4ad3bf9), a lot of bloat was introduced.
This commit moves bloat to the inside.

Config datatype got nice, typesafe params -- they allow to retrieve the
environment variables (or to get some defaults), while making sure that
the value will fit. However, the end result was not optimal, as end-user
had to perform e.g.:

    endpointRoot <- view (cWebEndpoint . unwrapDefaultableConf)

The last part -- the `unwrapDefaultableConf` -- is pure bloat that is
unnecessary.

To move this bloat somewhat inside:

- `DefaultableConfW` was created. Its sole responsibility is to switch
  between `DefaultableConf a b` and `a`.
- `Config` got a parameter that is passed to `DefaultableConfW`.
- `AppConfig` expects the unwrapped config, i.e. `Config 'DUnwrapped`.
- There is `unwrapCfg :: Config 'DWrapped -> Config 'DUnwrapped` that
  unwraps the `Config`.